+++ src/components/Cards/itemVaccine/item-vaccine-style.js
@@ -10,20 +10,19 @@ export const ContainerItemInformationVaccine = styled.View`
 `;
 
 export const ButtonInformationVaccine = styled.TouchableOpacity`
-  width: 370px;
   height: ${({ expanded }) => (expanded ? "91.25px" : "48.68px")};
+  display:flex;
+  flex-direction:row;
+  justify-content:center;
+  padding:8px 4px 4px 0px;
   background: #ffffff;
   border-bottom-width: 2px;
   border-bottom-color: rgba(76, 122, 187, 0.5);
 `;
 
 export const ViewInformationDate = styled.View`
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  width: 150px;
-  height: 35px;
-  margin-top: 5px;
+  width:50%;
+  padding:2px;
   border-right-width: 1px;
   border-left-width: 4px;
   border-left-color: ${({ colorBorderLeft }) => colorBorderLeft};
@@ -31,39 +30,33 @@ export const ViewInformationDate = styled.View`
 `;
 
 export const ContainerNameVaccine = styled.View`
-  flex: 1;
-  display: flex;
-  position: absolute;
-  top: 5px;
-  left: 170px;
-  width: 160px;
-  height: 22px;
-  background: rgba(76, 122, 187, 0.1);
-  border-radius: 10px;
+  display:flex;
+  flex-direction:row;
+  justify-content:space-around;
+  width: 50%;
+  height:25px;
 `;
 
 export const TextNameVaccine = styled.Text`
+  width:85%;
+  padding:2px 8px;
   font-weight: 500;
   font-size: 13px;
+  border-radius: 10px;
   color: #273047;
-  margin-left: 10px;
-  margin-top: 2px;
+  background: rgba(76, 122, 187, 0.1);
 `;
 
 export const TextDateInformation = styled.Text`
-  width: 120px;
   font-weight: 500;
   font-size: 12px;
   color: #273047;
-  margin-left: 2px;
 `;
 
 export const TextHourInformation = styled.Text`
-  width: 120px;
   font-weight: 500;
   font-size: 12px;
   color: #273047;
-  margin-left: 2px;
 `;
 
 export const MainContentVaccine = styled.Text`
@@ -98,16 +91,14 @@ export const ExpandedContainer = styled(ButtonInformationVaccine)`
 `;
 
 export const ViewItemHidden = styled.TouchableHighlight`
-  position: absolute;
-  right: 15px;
-  top: 5px;
-  width: 9.09px;
-  height: 18.26px;
+
+
 `;
 
 export const ImageHidden = styled.Image`
-  width: 18px;
-  height: 25px;
+ width:14px;
+ height:20px;
+ 
   transform: ${({ expanded }) =>
     expanded ? "rotate(180deg)" : "rotate(0deg)"};
 `;