+++ src/routes/Stack/index.tsx
@@ -1,19 +1,22 @@
 import React from "react";
-import SplashScreen from "../../screen/splashScreen/splashScreen";
-import LoginCadastro from "../../screen/loginCadastro/loginCadastro";
-import Login from "../../screen/login/login";
+import SplashScreen from "../../screen/splash-screen/splashScreen";
+
+
 import { propsNavigationStack } from "./Models";
-import PatientRecord from "../../screen/cadastro/cadastro";
-import TelaPrincipal from "../../screen/telaPrincipal/telaPrincipal";
+
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
-import PatientList from "../../screen/listaPacientes/lista-pacientes";
-import Patient from "../../screen/paciente/patient";
-import ProfissionalResponsavel from "../../screen/profissionalResponsável/profissionalResponsável";
-import MyVaccines from "../../screen/minhasVacinas/my-vaccines";
-import CodigoCelular from "../../screen/codigoCelular/codigo-celular";
-import RedefinirSenha from "../../screen/redefinirSenha/redefinirSenha";
-import RecuperacaoSenha from "../../screen/recuperacaoSenha/recuperacaoSenha";
-import TelaSucesso from "../../screen/telaSucesso/telaSucesso";
+import Patient from "../../screen/patient/patient";
+import PatientRecord from "../../screen/password-creation/cadastro";
+import TelaPrincipal from "../../screen/home/telaPrincipal";
+import ProfissionalResponsavel from "../../screen/responsibleDoctor/profissionalResponsável";
+import MyVaccines from "../../screen/myVaccines/my-vaccines";
+
+import RecuperacaoSenha from "../../screen/passwordRecovery/recuperacaoSenha";
+import TelaSucesso from "../../screen/passwordRecoverySuccess/telaSucesso";
+import PatientLogin from "../../screen/login/login";
+
+
+
 
 const { Navigator, Screen } =
   createNativeStackNavigator<propsNavigationStack>();
@@ -24,8 +27,8 @@ export default function () {
       initialRouteName="SplashScreen"
       screenOptions={{ headerShown: false }}
     >
-      <Screen name="LoginCadastro" component={LoginCadastro} />
-      <Screen name="Login" component={Login} />
+
+      <Screen name="Login" component={PatientLogin} />
       <Screen name="Cadastro" component={PatientRecord} />
       <Screen name="TelaPrincipal" component={TelaPrincipal} />
       <Screen name="Paciente" component={Patient} />
@@ -35,8 +38,8 @@ export default function () {
       />
       <Screen name="SplashScreen" component={SplashScreen} />
       <Screen name="MyVaccines" component={MyVaccines} />
-      <Screen name="CodePhone" component={CodigoCelular} />
-      <Screen name="RedefinirSenha" component={RedefinirSenha} />
+
+
       <Screen name="RecuperacaoSenha" component={RecuperacaoSenha} />
       <Screen name="TelaSucesso" component={TelaSucesso} />
       {/* <Screen name="TabRoutes" component={TabRoutes} /> */}