+++ src/screen/splash-screen/splashScreen.tsx
@@ -15,7 +15,7 @@ const SplashScreen: React.FC = () => {
   useEffect(() => {
 
     setTimeout(() => {
-      navigation.navigate("LoginCadastro");
+      navigation.navigate("Login");
     }, 3000);
   }, []);